rename `use`s
