+++ models/rust/src/3.1_RC2/files/station_information.rs
@@ -6,7 +6,7 @@ use napi_derive::napi;
 
 use serde::{Deserialize, Serialize};
 
-use crate::v3_0::types::*;
+use crate::v3_1_rc2::types::*;
 
 file_struct!(StationInformationFile, StationInformationData);
