+++ models/rust/src/lib.rs
@@ -2,4 +2,4 @@
 pub mod v3_0;
 
 #[path = "3.1_RC2/mod.rs"]
-pub mod v3_1_RC2;
+pub mod v3_1_rc2;
